pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "waterinfohub"
version = "0.1.0"
description = "Smart water market intelligence hub"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "alembic>=1.15.2",
  "apscheduler>=3.10.4",
  "beautifulsoup4>=4.12.3",
  "fastapi>=0.115.0",
  "feedparser>=6.0.11",
  "httpx>=0.28.1",
  "jinja2>=3.1.4",
  "pgvector>=0.3.6",
  "playwright>=1.52.0",
  "psycopg[binary]>=3.2.9",
  "pydantic>=2.11.0",
  "pydantic-settings>=2.8.1",
  "python-dateutil>=2.9.0.post0",
  "python-dotenv>=1.1.0",
  "pyyaml>=6.0.2",
  "sqlalchemy>=2.0.40",
  "tenacity>=9.1.2",
  "uvicorn>=0.34.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.5",
  "pytest-asyncio>=0.26.0",
  "ruff>=0.11.5",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["waterinfohub*"]

[tool.ruff]
line-length = 100
target-version = "py311"
